Edit: 사용자별로 todo 관련 CRUD 가능하도록 개선

+++ app/src/todos/repository.py
@@ -6,52 +6,57 @@ from typing import List
 from sqlalchemy.orm import Session
 from todos.models import Todo
 from todos.schemas import TodoCompletedStateToggle, TodoItem, TodoResponse
+from auth.models import User
 
 
 # READ: single todo
-def get_todo_by_id(todo_id: int, db: Session) -> Todo | None:
-    return db.query(Todo).filter(Todo.id == todo_id).first()
+def get_todo_by_id(todo_id: int, user_id: int, db: Session) -> Todo | None:
+    return db.query(Todo).filter(
+        Todo.id == todo_id,
+        Todo.user_id == user_id        
+        ).first()
 
 
 # READ: all todos
-def get_all_todos(db: Session) -> List[Todo] | None:
-    return db.query(Todo).all()
+def get_all_todos(user_id: int, db: Session) -> List[Todo] | None:
+    return db.query(Todo).filter(Todo.user_id == user_id).all()
 
 
 # CREATE: single todo
-def create_todo(todo: TodoItem, db: Session) -> Todo:
+def create_todo(todo: TodoItem, user_id: int, db: Session) -> Todo:
     new_todo = Todo(**todo.model_dump())
+    new_todo.user_id = user_id
     db.add(new_todo)
-    new_todo.task = todo.task
-    new_todo.due_date = todo.due_date
-    new_todo.is_done = False
     db.commit()
     db.refresh(new_todo)
     return new_todo
 
 
 # DELETE: all todos
-def remove_all_todos(db: Session) -> int:
-    deleted = db.query(Todo).delete()
+def remove_all_todos(user_id: int, db: Session) -> int:
+    deleted = db.query(Todo).filter(Todo.user_id == user_id).delete()
     db.commit()
     return deleted  # 삭제된 row 수
 
 
-# DELEte: single todo
-def remove_todo_by_id(todo_id: int, db: Session) -> int:
-    todo = get_todo_by_id(todo_id, db)
+# DELETE: single todo
+def remove_todo_by_id(todo_id: int, user_id: int, db: Session) -> int:
+    todo = get_todo_by_id(todo_id, user_id, db)
     if not todo:
-        return -1
-    deleted = db.query(Todo).filter(Todo.id == todo_id).delete()
+        return None
+    deleted = db.query(Todo).filter(
+        Todo.id == todo_id,
+        Todo.user_id == user_id
+        ).delete()
     db.commit()
     return deleted
 
 
 # UPDATE: todo
-def update_todo_by_id(todo_id: int, change: TodoItem, db: Session) -> Todo:
-    todo = get_todo_by_id(todo_id, db)
+def update_todo_by_id(todo_id: int, change: TodoItem, user_id: int, db: Session) -> Todo:
+    todo = get_todo_by_id(todo_id, user_id, db)
     if not todo:
-        return -1
+        return None
     todo.task = change.task
     todo.due_date = change.due_date
     db.commit()
@@ -60,11 +65,11 @@ def update_todo_by_id(todo_id: int, change: TodoItem, db: Session) -> Todo:
 
 
 def update_completed_state_by_id(
-    todo_id: int, state: TodoCompletedStateToggle, db: Session
+    todo_id: int, state: TodoCompletedStateToggle, user_id: int, db: Session
 ) -> Todo:
-    todo = get_todo_by_id(todo_id, db)
+    todo = get_todo_by_id(todo_id, user_id, db)
     if not todo:
-        return -1
+        return None
     todo.is_done = state.is_done
     db.commit()
     db.refresh(todo)
